added redux folder and file structure
