+++ note/src/components/NoteDelete.js
