+++ note/src/index.js
@@ -4,11 +4,7 @@ import './index.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 
-import { BrowserRouter } from 'react-router-dom'
-
 ReactDOM.render(
-    (<BrowserRouter>
-        <App />
-    </BrowserRouter>),
+        <App />,
     document.getElementById('root'));
 registerServiceWorker();
