+++ note/src/noteData.js
@@ -2,38 +2,47 @@ const noteData = [
     {
         title: "Note Title",
         text: "Morbi pellentesque euismod venenatis. Nulla ut nibh nunc. Phasellus diam metus, blandit ac purus a, efficitur mollis …",
+        id: 0,
     },
     {
         title: "Note Title",
         text: "Morbi pellentesque euismod venenatis. Nulla ut nibh nunc. Phasellus diam metus, blandit ac purus a, efficitur mollis …",
+        id: 1,
     },
     {
         title: "Note Title",
         text: "Morbi pellentesque euismod venenatis. Nulla ut nibh nunc. Phasellus diam metus, blandit ac purus a, efficitur mollis …",
+        id: 2,
     },
     {
         title: "Note Title",
         text: "Morbi pellentesque euismod venenatis. Nulla ut nibh nunc. Phasellus diam metus, blandit ac purus a, efficitur mollis …",
+        id: 3,
     },
     {
         title: "Note Title",
         text: "Morbi pellentesque euismod venenatis. Nulla ut nibh nunc. Phasellus diam metus, blandit ac purus a, efficitur mollis …",
+        id: 4,
     },
     {
         title: "Note Title",
         text: "Morbi pellentesque euismod venenatis. Nulla ut nibh nunc. Phasellus diam metus, blandit ac purus a, efficitur mollis …",
+        id: 5,
     },
     {
         title: "Note Title",
         text: "Morbi pellentesque euismod venenatis. Nulla ut nibh nunc. Phasellus diam metus, blandit ac purus a, efficitur mollis …",
+        id: 6,
     },
     {
         title: "Note Title",
         text: "Morbi pellentesque euismod venenatis. Nulla ut nibh nunc. Phasellus diam metus, blandit ac purus a, efficitur mollis …",
+        id: 7,
     },
     {
         title: "Note Title",
         text: "Morbi pellentesque euismod venenatis. Nulla ut nibh nunc. Phasellus diam metus, blandit ac purus a, efficitur mollis …",
+        id: 8,
     },
 ];
